+++ CHANGELOG.md
@@ -9,10 +9,12 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/) and this p
 ### Added:
 - Resource links to empty state Volumes landing page #9065
 - Resource links to empty state Firewalls landing page #9078
+- Resource links to empty state Domains landing page #9092
 - Ability download DNS zone file #9075
 
 ### Changed:
 - Removed MongoDB reference from ClusterControl description #9081
+- Highlighted Marketplace apps and button card height on empty state Linodes landing page #9083
 
 ### Fixed:
 - Ability to search Linodes by IPv6 #9073
@@ -25,6 +27,9 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/) and this p
 - MUI v5 Migration - `Components > TableBody` #9082
 - MUI v5 Migration - `Components > TableSortCell` #9082
 - MUI v5 Migration - `Components > TableRow` #9082
+- MUI v5 Migration - `Components > DownloadCSV` #9084
+- MUI v5 Migration - `Components > Notice` #9094
+- MUI v5 Migration - `Components > CheckoutSummary` #9100
 - React Query for Linodes Landing #9062
 
 ## [2023-05-01] - v1.92.0